Throw away old tables
