--- src/Components/Table/Table.scss
@@ -1,175 +1,0 @@
-@import '../../assets/styles/base/variables';
-@import '../../assets/styles/base/mixins/media';
-
-.table {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-
-  &__desktop {
-    @include respond-below(sm) {
-      display: none;
-    }
-  }
-
-  &__mobile {
-    @include respond-above(sm) {
-      display: none;
-    }
-    @include respond-below(sm) {
-      padding: 5px;
-      display: flex;
-      flex-direction: column;
-      background: #ffffff;
-      border-radius: 5px 5px 0 0;
-    }
-
-    p {
-      align-self: center;
-      font-size: $smaller-size;
-      padding: 5px 0 5px 0;
-      color: $secondary-text-color;
-    }
-  }
-
-  &__mobile-row {
-    justify-content: space-between;
-    flex-wrap: wrap;
-    font-size: $small-size;
-    width: 100%;
-    margin-bottom: 5px;
-    &:last-child {
-      margin-bottom: 0px;
-    }
-  }
-
-  &__header {
-    width: 100%;
-    display: flex;
-    justify-content: flex-start;
-    align-items: center;
-    padding: 20px 40px 20px 40px;
-
-    .cell {
-      font-style: normal;
-      font-weight: $medium-weight;
-      font-size: $smaller-size;
-      line-height: 18px;
-      color: #AAAAAA;
-      text-overflow: ellipsis;
-      overflow: hidden;
-      white-space: nowrap;
-    }
-  }
-
-  &__body {
-    height: 86%;
-    width: 100%;
-    padding: 20px;
-    background: #ffffff;
-    display: flex;
-    flex-direction: column;
-    overflow-x: scroll;
-
-    .empty {
-      width: 100%;
-      height: 100%;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      .no-data {
-        font-style: normal;
-        font-weight: $normal-weight;
-        font-size: $large-size;
-        line-height: 21px;
-        color: $names-text-color;
-      }
-    }
-
-    .row {
-      background: #FFFFFF;
-      border: 1px solid #ECECF2;
-      box-sizing: border-box;
-      border-radius: 12px;
-      padding: 20px;
-      display: flex;
-      align-items: center;
-      justify-content: flex-start;
-      margin-bottom: 15px;
-
-      .cell {
-        padding-right: 0.2em;
-        font-style: normal;
-        font-weight: $medium-weight;
-        font-size: $accent-size;
-        line-height: 24px;
-        color: $main-text-color;
-        text-overflow: ellipsis;
-        overflow: hidden;
-        white-space: nowrap;
-      }
-    }
-  }
-
-  &__footer {
-    display: flex;
-    justify-content: space-between;
-    border-top: 1px solid #EEEEEE;
-    align-items: center;
-    background: #FFFFFF;
-    border-bottom-left-radius: 10px;
-    border-bottom-right-radius: 10px;
-    padding: 20px;
-
-    &__pagination {
-    }
-
-    &__button {
-      background: rgba(129, 129, 165, 0.1) !important;
-      border-radius: 6px;
-      cursor: pointer;
-
-      p {
-        font-family: Lato;
-        font-style: normal;
-        font-weight: $bold-weight;
-        font-size: $smaller-size;
-        line-height: 14px;
-        text-transform: uppercase;
-        color: $inputs-labels-text-color;
-      }
-    }
-
-    &__button.next {
-      padding: 11px 35px 11px 15px;
-      text-align: right;
-    }
-
-    &__button.prev {
-      padding: 11px 15px 11px 35px;
-      text-align: left;
-    }
-  }
-
-  .MuiPaginationItem-root.MuiPaginationItem-ellipsis {
-    display: none;
-  }
-
-  .MuiButtonBase-root.MuiPaginationItem-root.MuiPaginationItem-page.MuiPaginationItem-outlined.MuiPaginationItem-rounded {
-    border: none;
-    background: #F5F5FA;
-    border-radius: 6px;
-    width: 36px;
-    height: 36px;
-    color: $inputs-labels-text-color;
-  }
-
-  .MuiButtonBase-root.MuiPaginationItem-root.MuiPaginationItem-page.MuiPaginationItem-outlined.MuiPaginationItem-rounded.Mui-selected {
-    background: linear-gradient(0deg, rgba(38, 34, 98, 0.4), rgba(38, 34, 98, 0.4)), #FFFFFF;
-    color: #FFFFFF;
-    position: relative;
-  }
-
-}
-
